requirements.txt
python-dotenv
requests
pandas
pytest
python-3.12.8
worker: PYTHONPATH=. python -m app.main
